Update Dropbox SDK as requested by Dropbox in email
We are reaching out to you because we have discovered an issue with the
Dropbox API v2 Python SDK that your app is currently using. Although the
issue does not break any existing behavior, it limits our ability to
make non-breaking changes to our API endpoints going forward.

For that reason, we ask that you update your app to the latest version
of the SDK within the next two months. This bug was fixed specifically
in v8.0.0.